Start of 3rd person camera controls and player control integration
player isn't animating for movement because the variables are different, and aiming isn't working like before. There are also issues with rotation because of the way I've set this up but that's not a major concern for now. Need to make aiming work in 3rd person, probably going to make it just line cast using camera forward vector